backend/requirements.txt
flask==2.0.3
werkzeug==2.0.3
psycopg2-binary==2.9.6
pika==1.3.1
flask-cors==3.0.10
gunicorn==20.1.0